feat: rip gamekeys from useroptions from src of /keys
when loading https://www.humblebundle.com/home/keys the src contains
some user options as JSON, one of which is a list of all gamekeys (which I
believe are all the orders made from that account). Once the page is
loaded, this data seems not to exist anymore, I assumed it's removed by
some kind of javascript.